% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{rolling_window}
\alias{rolling_window}
\title{Make rolling windows}
\usage{
rolling_window(x, l, u)
}
\arguments{
\item{x}{A numeric vector.}

\item{l}{An integer specifying how far away the lower bound should be.}

\item{u}{An integer specifying how far away the upper bound should be.}
}
\value{
A Data frame with the rolling window values of the mean, median, min, and max.
}
\description{
Computes rolling mean, median, min, max over a specified window, over a vector.
}
