% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2df.R
\name{ts2df}
\alias{ts2df}
\title{For those of us who loathe time-series objects.}
\usage{
ts2df(ts)
}
\arguments{
\item{ts}{A time-series object.}
}
\value{
A Data frame with two columns: date and value.
}
\description{
Converts a tis object into a data frame.
}
